Implement simple ad management javascript library
In version 1.0 there was no proper separation of the frontend and backend
functionality. The backend was in charge to generate the scripts for the
frontend.

So a new javascript library is implemented which handles only the frontend
logic. It loads the current ad and displays it based on the configured
timeout. An AdTracker object deals with the tracking of frontend actions,
it sends an AJAX request to a configurable endpoint to the server.