Improve support for custom renderers
Each ad has a renderer. This is a simple object which shows/hides the ad.
Custom renderers can be written (ie. display ad with fancybox or in other
ways) and we must support them. So we must allow the client code to specify
the script name which contains the renderer object.

The ad JSON data must contain the renderer configuration like this:
{
 ...
 "renderer": {name: "SimpleRenderer", script: "assets/vendor/konekt/kampaign/js/renderers/simple.js"}
}